mysite_django_arduino/requirements.txt
django
pyserial
python-dotenv
